backend/requirements.txt
aixplain
aiosignal
annotated-types
anyio
attrs
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
dataclasses-json
filelock
Flask
Flask-Cors
fpdf
frozenlist
future
h11
httpcore
httpx
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
MarkupSafe
marshmallow
multidict
mypy-extensions
numpy
packaging
pycparser
pydantic
pydantic_core
PyMuPDF
PyYAML
python-docx
python-pptx
regex
requests
six
sniffio
SQLAlchemy
tenacity
tqdm
typing-inspect
typing_extensions
urllib3
Werkzeug
yarl